% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_fit.R
\name{get.mediation.initdata}
\alias{get.mediation.initdata}
\title{get.mediation.initdata}
\usage{
get.mediation.initdata(data, covariates, sl)
}
\description{
helper function called within SDE_tmle to perform initial regressions for TMLE and
estimating equation (one step estimator).
}
