% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_fitNT.R
\name{get.mediation.initdataNT}
\alias{get.mediation.initdataNT}
\title{get.mediation.initdataNT}
\usage{
get.mediation.initdataNT(data, covariates, sl)
}
\description{
helper function called within SDE_tmleNT to perform initial regressions for TMLE and
estimating equation (one step estimator).
}
