% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gstarM.R
\name{get_gstarM}
\alias{get_gstarM}
\title{get_gstarM}
\usage{
get_gstarM(data, sl, V = 10, covariates, transport)
}
\description{
helper function called within SDE_tmle to compute the probability of mediator M
under the stochastic intervention
}
