% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_steps.R
\name{mediation.step1}
\alias{mediation.step1}
\title{mediation.step1}
\usage{
mediation.step1(initdata, Y_preds, data, gstarM_astar, a, transport)
}
\description{
helper function called within SDE_tmle to perform TMLE update for the first regression
}
