% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_steps.R
\name{mediation.step2}
\alias{mediation.step2}
\title{mediation.step2}
\usage{
mediation.step2(data, sl, Qstar_M, Qstar_Mg, covariates_QZ, Hm, A_ps, a,
  tmle = TRUE, EE = FALSE, transport)
}
\description{
helper function called within SDE_tmle to perform TMLE update for the 2nd regression
and also performs estimating equation (one step estimator)
}
