% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDE_tmle.R
\name{SDE_tmle}
\alias{SDE_tmle}
\title{SDE_tmle}
\usage{
SDE_tmle(data, sl, V = 10, covariates, truncate = list(lower = 1e-04, upper
  = 0.9999), alpha = 0.05, glm_only = FALSE)
}
\arguments{
\item{data, }{data.frame of variables in time ordering from left to right. Make sure if using a
continuous outcome that it is scaled to be between 0 and 1 via (Y - minY)/(maxY - minY).  The mediator
M, intermediate confounder Z, treatment A,and site S are all binaries and must be named as such.}

\item{sl}{the sl3 superlearner defined, see sl3 documentation for defining a superlearner
and the example below}

\item{V}{number of folds for cross-validation (fixed to 10 for now)}

\item{covariates, }{dataframe of covariates for each necessary model, going backwards from the
outcome.}

\item{truncate, }{a list with elements lower and upper to truncate the various p-scores
not functional at present}

\item{alpha}{= type I error rate}

\item{glm_only}{= fit using main terms glm}

\item{a, }{the treatment intervention of interest, either 0 or 1}

\item{a_star, }{the treatment intervention of the stochastic model for Mediator, M, either 0 or 1.}

\item{CI}{confidence intervals for tmle, one step estimator, iptw estimators for stochastic
direct and indirect effects}

\item{IC}{dataframe of influence curve approximations for tmle, the one step estimator,
and iptw estimator for both stochastic direct and indirect effects}

\item{gcomp_ests}{stochastic direct and indirect effects for gcomp, IC inference is
generally not valid here so no confidence interval offered.}

\item{fluctuation_eps}{The tmle here uses the clever covariate as a weight in an intercept
logistic regression model. This returns the intercepts for the three parameters defined as
means, respectively, under the stochastic interventions astar = a = 0, astar = a = 1 and
astar = 0 with a = 1.}
}
\value{
a list with the following elements
}
\description{
computes the sequential regression, targeted maximum likelihood estimate
for the stochastic direct effect and stochastic indirect effect when the outcome and
mediator model are only available on site 1 (S = 1).  This is a data adaptive parameter
as the stochastic direct effect has a model for the mediator is determined on the data for
site 1. We note, we are working on future generalizations to allow the mediator to be present
for both sites and also to allow specification of separate superlearners for all models that are
fitted.
}
\examples{
data(data_example)
head(data_example)

covariates = list(covariates_S = c("W1","W2"),
                  covariates_A = c("S","W1", "W2"),
                  covariates_Z = c("S","A","W1","W2"),
                  covariates_M = c("Z","W1","W2"),
                  covariates_Y = c("M","Z","W1","W2"),
                  covariates_QZ = c("S","W1","W2"))

# make the superlearner using sl3
lglm = make_learner(Lrnr_glm)
lmean = make_learner(Lrnr_mean)
# lxgboost = make_learner(Lrnr_xgboost)
lrnr_stack = make_learner(Stack, list(lglm, lmean))
metalearner = make_learner(Lrnr_nnls)

# define the superlearner
sl <- Lrnr_sl$new(learners = lrnr_stack,
                  metalearner = metalearner)

# takes about 9 seconds
time = proc.time()
result = SDE_tmle(data_example, sl = sl, covariates = covariates, 
                   truncate = list(lower =.0001, upper = .9999), alpha = .05, glm_only = FALSE)
proc.time() - time 

result$CI
head(result$IC)
result$gcomp_ests
result$fluctuation_eps

}
