% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDE_tmleNT.R
\name{SDE_tmleNT}
\alias{SDE_tmleNT}
\title{SDE_tmleNT}
\usage{
SDE_tmleNT(data, sl, V = 10, covariates, truncate = list(lower = 1e-04,
  upper = 0.9999), alpha = 0.05)
}
\description{
Called within SDE_tmle to compute stochastic direct and indirect effects when no
transporting is desired
}
